Fichier test complet + quelques broutilles + commentaires sur acc

+++ DMA.cpp
@@ -27,7 +27,7 @@ int main()
     //On envoie un tableau noir
     for(unsigned int i = 0; i < 4; i++)
         for(unsigned int j = 1; j < 4; j+=2)
-            DMA.writeSourceInteger(blackImg[i][j-1] << 24 | whiteImg[i][j-1] << 16 | blackImg[i][j] << 8 | whiteImg[i][j]);
+            DMA.writeSourceInteger((unsigned int) (blackImg[i][j-1] << 24 | whiteImg[i][j-1] << 16 | blackImg[i][j] << 8 | whiteImg[i][j]));
 
     //Configuration de l'interruption
     DMA.setInterrupt(true,true,0xFF);
